+++ CHANGELOG.md
@@ -10,8 +10,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [How to upgrade to the latest version!](https://unicorn-binance-local-depth-cache.docs.lucit.tech/readme.html#installation-and-upgrade)
 
 ## 2.7.0.dev (development stage/unreleased/unstable)
+
+## 2.8.0
 ### Changed
-- `get_list_of_depth_caches()` is deprecated, use `get_list_of_depthcaches()` instead!
+- `get_list_of_depth_caches()` is deprecated, use `get_list_of_depthcaches()` instead! From now on, only active 
+  DepthCaches are returned!
+- `stop_depthcache()` removes asks and bids, but the rest of the DepthCache remains.
 
 ## 2.7.0
 ### Changed